refactor: move common datatypes into separate files
